requirements.txt
tornado
tensorflow-gpu
psycopg2-binary
selenium
bs4
requests
scipy
pillow

